lint: rename files
