+++ src/commands/install-buf.ts
