+++ src/commands/show-output.ts
