+++ test/unit/commands/buf-generate.test.ts
@@ -1,22 +1,20 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 
+import * as vscode from "vscode";
 import * as assert from "assert";
 import * as sinon from "sinon";
-import * as vscode from "vscode";
-import * as cmds from "../../../src/commands";
 import * as util from "../../../src/util";
-
+import { bufGenerate } from "../../../src/commands/buf-generate";
 import { BufContext } from "../../../src/context";
-import { MockExtensionContext } from "../../mocks/MockContext";
+import { MockExtensionContext } from "../../mocks/mock-context";
 
-suite("commands.loadBufModules", () => {
-  vscode.window.showInformationMessage("Start all loadBufModules tests.");
+suite("commands.bufGenerate", () => {
+  vscode.window.showInformationMessage("Start all bufGenerate tests.");
 
   let sandbox: sinon.SinonSandbox;
   let execFileStub: sinon.SinonStub;
   let logErrorStub: sinon.SinonStub;
   let logInfoStub: sinon.SinonStub;
-  let createFileSystemWatcherSpy: any;
 
   let ctx: any;
   let bufCtx: BufContext;
@@ -58,12 +56,7 @@ suite("commands.loadBufModules", () => {
         } as unknown as vscode.Disposable;
       });
 
-    createFileSystemWatcherSpy = sandbox.spy(
-      vscode.workspace,
-      "createFileSystemWatcher"
-    );
-
-    cmds.loadBufModules.register(ctx, bufCtx);
+    bufGenerate.register(ctx, bufCtx);
   });
 
   teardown(() => {
@@ -71,38 +64,52 @@ suite("commands.loadBufModules", () => {
     bufCtx.dispose();
   });
 
-  test("construction sets up file watchers", async () => {
-    assert.strictEqual(createFileSystemWatcherSpy.calledOnce, true);
-    assert.strictEqual(
-      createFileSystemWatcherSpy.getCall(0).args[0],
-      "**/buf.yaml"
-    );
-  });
-
-  test("logs an error and does nothing if buf is not installed", async () => {
+  test("should log an error if buf is not installed", async () => {
     bufCtx.buf = undefined;
 
     await cmdCallback();
 
     assert.strictEqual(logErrorStub.calledOnce, true);
-    assert.strictEqual(execFileStub.notCalled, true);
   });
 
-  test("runs 'buf ls-files'", async () => {
+  test("should call 'buf generate'", async () => {
     bufCtx.buf = { path: "/path/to/buf", version: "1.0.0" } as any;
-
-    execFileStub.resolves({
-      stdout: `{"path":"external/google/type/latlng.proto","import_path":"google/type/latlng.proto","module":"google","commit":"","is_import":false}`,
-      stderr: "",
-    });
+    execFileStub.resolves({ stdout: "Generated successfully", stderr: "" });
 
     await cmdCallback();
 
     assert.strictEqual(execFileStub.calledOnce, true);
-    assert.strictEqual(bufCtx.bufFiles.size, 1);
+    assert.deepStrictEqual(execFileStub.args[0], [
+      "/path/to/buf",
+      ["generate"],
+      { cwd: vscode.workspace.rootPath },
+    ]);
+    assert.strictEqual(logInfoStub.calledWith("Generated successfully"), true);
+  });
+
+  test("should throw an error if anything is written to stderr", async () => {
+    bufCtx.buf = { path: "/path/to/buf", version: "1.0.0" } as any;
+    execFileStub.resolves({ stdout: "", stderr: "Error occurred" });
+
+    await cmdCallback();
+
+    assert.strictEqual(logErrorStub.calledOnce, true);
     assert.strictEqual(
-      bufCtx.bufFiles.get("external/google/type/latlng.proto")?.module,
-      "google"
+      logErrorStub.calledWith("Error generating buf: Error occurred"),
+      true
+    );
+  });
+
+  test("should throw an error if executing buf throws an error", async () => {
+    bufCtx.buf = { path: "/path/to/buf", version: "1.0.0" } as any;
+    execFileStub.rejects(new Error("Execution failed"));
+
+    await cmdCallback();
+
+    assert.strictEqual(logErrorStub.calledOnce, true);
+    assert.strictEqual(
+      logErrorStub.calledWith("Error generating buf: Execution failed"),
+      true
     );
   });
 });
